+++ src/main/java/org/hangman/ui/Gallows.java
@@ -1,4 +1,4 @@
-package org.hangman;
+package org.hangman.ui;
 
 public class Gallows {
