+++ src/main/java/org/hangman/ui/PlayerInteraction.java
@@ -1,4 +1,4 @@
-package org.hangman;
+package org.hangman.ui;
 
 import java.util.Scanner;
